Extract the code from TaskController to TaskRepository
