+++ app/src/main/java/com/codesoom/assignment/controllers/TaskController.java
@@ -1,6 +1,7 @@
 package com.codesoom.assignment.controllers;
 
 import com.codesoom.assignment.models.Task;
+import com.codesoom.assignment.models.TaskRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -25,68 +26,35 @@ import java.util.Optional;
 @RequestMapping("/tasks")
 @CrossOrigin
 public class TaskController {
-    private final List<Map<String, Task>> tasks = new ArrayList<>();
-    private final Map<String, Task> taskMap = new HashMap<>();
-    private Long newId = 0L;
+
+    private final TaskRepository taskRepository;
+
+    public TaskController(TaskRepository taskRepository) {
+        this.taskRepository = taskRepository;
+    }
 
     @GetMapping
     public List<Map<String, Task>> getTasks() {
-        return tasks;
+        return taskRepository.getTasks();
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<Task> getTask(@PathVariable("id") Long id) {
-        Optional<Task> task = findTask(id);
-        if (task.isEmpty()) {
-            return new ResponseEntity(HttpStatus.NOT_FOUND);
-        }
-        return new ResponseEntity(task, HttpStatus.OK);
+        return taskRepository.getTask(id);
     }
 
     @PostMapping
     public ResponseEntity<Task> createTask(@RequestBody Task task) {
-        task.setId(generateId());
-        taskMap.put(task.getId() + "", task);
-        if (tasks.isEmpty()) {
-            tasks.add(taskMap);
-        }
-
-        return new ResponseEntity(task, HttpStatus.CREATED);
+        return taskRepository.createTask(task);
     }
 
     @RequestMapping(path = "/{id}", method = {RequestMethod.PATCH, RequestMethod.PUT})
     public ResponseEntity<Task> updateTask(@PathVariable("id") Long id, @RequestBody Task task) {
-        Optional<Task> findtask = findTask(id);
-        if (findtask.isEmpty()) {
-            return new ResponseEntity(HttpStatus.NOT_FOUND);
-        }
-        findtask.get().setTitle(task.getTitle());
-
-        return new ResponseEntity(findtask, HttpStatus.OK);
+        return taskRepository.updateTask(id, task);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity deleteTask(@PathVariable("id") Long id) {
-        Optional<Task> findtask = findTask(id);
-        if (findtask.isEmpty()) {
-            return new ResponseEntity(HttpStatus.NOT_FOUND);
-        }
-        taskMap.remove(String.valueOf(id));
-
-        return new ResponseEntity(HttpStatus.NO_CONTENT);
-    }
-
-    private Long generateId() {
-        return ++newId;
-    }
-
-    private Optional<Task> findTask(Long id) {
-        Optional<Task> task = Optional.empty();
-        Task findTask = taskMap.get(String.valueOf(id));
-
-        if (findTask == null) {
-            return task;
-        }
-        return task.of(findTask);
+        return  taskRepository.deleteTask(id);
     }
 }
